Move `polyRootsImpl` back into `Autodiff.ts`
