+++ packages/core/src/engine/Optimizer.ts
@@ -1,5 +1,4 @@
 import consola from "consola";
-import { EigenvalueDecomposition, Matrix } from "ml-matrix";
 import { zip2 } from "../utils/Util.js";
 import * as lbfgs from "./Lbfgs.js";
 
@@ -458,32 +457,4 @@ const containsNaN = (numberList: Float64Array): boolean => {
     if (Number.isNaN(n)) return true;
   }
   return false;
-};
-
-/**
- * Replaces the contents of `v` with the roots of the monic polynomial whose
- * degree is the length of the vector and whose coefficient with a given degree
- * is the element of the vector at that index. Any root with a nonzero imaginary
- * component is replaced with `NaN`.
- */
-export const polyRoots = (v: Float64Array): void => {
-  const n = v.length;
-  // https://en.wikipedia.org/wiki/Companion_matrix
-  const m = Matrix.zeros(n, n);
-  for (let i = 0; i + 1 < n; i++) {
-    m.set(i + 1, i, 1);
-    m.set(i, n - 1, -v[i]);
-  }
-  m.set(n - 1, n - 1, -v[n - 1]);
-
-  // the characteristic polynomial of the companion matrix is equal to the
-  // original polynomial, so by finding the eigenvalues of the companion matrix,
-  // we get the roots of its characteristic polynomial and thus of the original
-  // polynomial
-  const r = new EigenvalueDecomposition(m);
-  for (let i = 0; i < n; i++) {
-    // as mentioned in the `polyRoots` docstring in `engine/AutodiffFunctions`,
-    // we discard any non-real root and replace with `NaN`
-    v[i] = r.imaginaryEigenvalues[i] === 0 ? r.realEigenvalues[i] : NaN;
-  }
 };
